+++ modules/products-list-component/src/products-list.module.ts
@@ -1,3 +1,4 @@
+import { RouterModule } from '@angular/router';
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { ProductsListComponent } from './products-list.component';
@@ -8,7 +9,13 @@ import { TagsSearchModule } from '@nodeart/tags-search-component';
  * Display products from specific category
  */
 @NgModule({
-    imports: [ProductServiceModule, BrowserModule, AttributesSearchModule, TagsSearchModule],
+    imports: [
+        ProductServiceModule, 
+        BrowserModule, 
+        AttributesSearchModule, 
+        TagsSearchModule,
+        RouterModule
+    ],
     exports: [ProductsListComponent],
     declarations: [ProductsListComponent],
     providers: [],